Fixed a bug where message sends to all subscriptions regardless
Messages were sending to all subscriptions regardless of topic. This
was because we were taking a cartesian product over (topic, message,
sub) and only checking the topic and message topic arns were
equal.

In the process, I've added a bit more loggin. Maybe in the future,
I'll put this behind a verbose flag.